Add files via upload
At the time of creating the application, I was not proficient in using Github. However, the application I created is an IT service management solution that enables organizations to track their assets, manage support tickets, and provide automated help desk support. Users can create, track, and manage tickets, which follow a hierarchical approval process where tickets are escalated to higher-level managers for approval.

In addition, the application includes reports and dashboards that allow managers to monitor the performance of their services and identify areas for improvemen